+++ data/Store.js
@@ -69,13 +69,15 @@ export class Store {
      *      StoreFilter config to create.
      * @param {boolean} [c.loadRootAsSummary] - true to treat the root node in hierarchical data as
      *      the summary record.
+     * @param {Object[]} [c.data] - source data to load
      */
     constructor({
         fields,
         idSpec = 'id',
         processRawData = null,
         filter = null,
-        loadRootAsSummary = false
+        loadRootAsSummary = false,
+        data
     }) {
         this.fields = this.parseFields(fields);
         this.idSpec = isString(idSpec) ? (data) => data[idSpec] : idSpec;
@@ -85,6 +87,8 @@ export class Store {
 
         this.resetRecords();
         this.setFilter(filter);
+
+        if (data) this.loadData(data);
     }
 
     /** Remove all records from the store. Equivalent to calling `loadData([])`. */
@@ -161,8 +165,6 @@ export class Store {
      */
     @action
     updateData(rawData) {
-        const changeLog = {};
-
         // Build a transaction object out of a flat list of adds and updates
         let rawTransaction = null;
         if (isArray(rawData)) {