+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     properties of the `AgGrid` component
 * The `Store` constructor now accepts a `data` argument to load data at initialization.
 * Added a `PinPad` component for streamlined handling of PIN entry on mobile devices.
+* The Grid context menu now contains an `Autosize Columns` option by default.
+* `GridModel` now exposes an `autoSizeColumns` method.
 * `Column` and `ColumnGroup` now support React elements for `headerName`.
 
 ### 🐞 Bug Fixes