+++ CHANGELOG.md
@@ -13,6 +13,11 @@
 * `DataViewModel` now accepts `groupBy` and `groupedItemHeight` params to enable grouped `DataView`.
   * `DataViewModel` also accepts other styling properties of the `AgGrid` component: `sizingMode`,
     `showHover`, `rowBorders`, and `stripeRows`.
+* The `Store` constructor now accepts a `data` argument to load data at initialization.
+* Added a `PinPad` component for streamlined handling of PIN entry on mobile devices.
+* The Grid context menu now contains an `Autosize Columns` option by default.
+* `GridModel` now exposes an `autoSizeColumns` method.
+* `Column` and `ColumnGroup` now support React elements for `headerName`.
 
 ### 🐞 Bug Fixes