+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   directly by applications that wish to customize the default row heights globally on init
   * 💥 Note that these were previously exported as similar constants from AgGrid.js. This would be
     a breaking change for any apps that imported the old objects directly (considered unlikely).
+* `FormFields` now check that they are contained in a Hoist `Form`.
 
 ### 📚 Libraries