+++ CHANGELOG.md
@@ -10,8 +10,14 @@
   * Each of these sizing modes has their own css variables for applications to override as needed
   * Header and row heights are configurable for each via the `HEADER_HEIGHTS` or `ROW_HEIGHTS`
     properties of the `AgGrid` component
+* The `Store` constructor now accepts a `data` argument to load data at initialization.
+* Added a `PinPad` component for streamlined handling of PIN entry on mobile devices.
+* The Grid context menu now contains an `Autosize Columns` option by default.
+* `GridModel` now exposes an `autoSizeColumns` method.
+* `Column` and `ColumnGroup` now support React elements for `headerName`.
 * The `xh/hoist/data/cube` package has been modified substantially to better integrate with the core
     data package and support observable "Views".  See documentation on `Cube` for more information.
+
 
 ### 🐞 Bug Fixes